Support terraform 0.15.
In terraform 0.14 outputs that are senstive need to be marked with
`sensitive = true` or the run will fail. This also means that
we need to use terraform output -json and parse the output or
we won't get the real value.

In terraform 0.15 `terraform destroy -force` was changed to
`terraform destroy -auto-approve`.