Correction to Lara cheat sheet
The SSH key pair generated for the RCE needs to have a public key that's short enough for the RCE. A long public key, such as an RSA public key generated with ssh-keygen, will get truncated in the RCE and cause the command to fail. Using an ed25519 key pair solves this issue.