Fix container build on non-arm64 hosts.
This changes the Dockerfile to be based on the upstream docker container
rather then pulling the executable down ourselves. Before we where
accidentally fetching the arm64 version, which of course won't work on
amd64 hosts.

I'm not positive, but I believe doing it this way is platform independent,
if someone has a arm64 mac and can test this that would be appreciated.
In any case this resolves the issue running the container on my amd host.

Resolves #122