Update README.md
Redesigned the scenarios so they are sorted by "Easy" "Medium" and "Hard". I also made it so they are drop down menus for users to find the correct difficulty easier. 

The old way of just having them all randomly together; with the size and difficulty in the name was confusing.